refactor(product): repository 이동
