+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.application;
 
 import kitchenpos.menus.application.MenuService;
-import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import kitchenpos.products.dto.ProductRequest;
 import kitchenpos.profanity.ProfanityClient;
 import kitchenpos.products.tobe.domain.Product;
